Add Invoker and Function and Loader features
To invoke unit-funciton in view or controller context.
Invoker is to be included in ActionView::Base and ActionController::Base,
and provide a `#invoke` method to invoke a funciton defined in an unit.

In an execution of `#invoke`,
Options parses its arguments and provides easy interfaces to invoker,
and FunctionFinder finds a matching function for it.

Loader is created to load units and cache loaded units,
but these feature is not implemented yet for now.